Latest solita

+++ programs/dark-bonds/src/common/constants.rs
@@ -1,4 +1,4 @@
-pub const SECONDS_YEAR: f64 = 31536000.0;
+pub const SECONDS_YEAR: i64 = 31536000;
 pub const SECONDS_IN_A_DAY: i64 = 86400;
 pub const PURCHASE_CUT: f64 = 5.5; // equivalent to 5%
